+++ src/onegov.swissvotes/onegov/swissvotes/models/vote.py
@@ -361,3 +361,13 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     @observes('files')
     def files_observer(self, files):
         self.vectorize_files()
+
+    def get_file(self, name):
+        """ Returns the requested localized file, falls back to the default
+        locale.
+
+        """
+        fallback = SwissVote.__dict__.get(name).__get_by_locale__(
+            self, self.session_manager.default_locale
+        )
+        return getattr(self, name, None) or fallback